+++ test/components/category_test.dart
@@ -15,6 +15,54 @@ void main() {
       expect(find.byType(GestureDetector), findsNothing);
     });
 
+    testWidgets('Keep hashtag', (tester) async {
+      final tenorCategoryTest = TenorCategory(
+        name: '#test',
+        searchTerm: 'test search term',
+        path: 'path/to/category',
+        image: 'https://flyclops.com/images/logo.png',
+      );
+
+      await tester.pumpWidget(
+        Directionality(
+          textDirection: TextDirection.ltr,
+          child: TenorCategoryWidget(
+            category: tenorCategoryTest,
+            style: const TenorCategoryStyle(
+              stripHashtag: false,
+            ),
+          ),
+        ),
+      );
+
+      expect(find.text('test'), findsNothing);
+      expect(find.text('#test'), findsOneWidget);
+    });
+
+    testWidgets('Strip hashtag', (tester) async {
+      final tenorCategoryTest = TenorCategory(
+        name: '#test',
+        searchTerm: 'test search term',
+        path: 'path/to/category',
+        image: 'https://flyclops.com/images/logo.png',
+      );
+
+      await tester.pumpWidget(
+        Directionality(
+          textDirection: TextDirection.ltr,
+          child: TenorCategoryWidget(
+            category: tenorCategoryTest,
+            style: const TenorCategoryStyle(
+              stripHashtag: true,
+            ),
+          ),
+        ),
+      );
+
+      expect(find.text('test'), findsOneWidget);
+      expect(find.text('#test'), findsNothing);
+    });
+
     testWidgets('Make sure it renders and is tappable', (tester) async {
       // track if the tap was tapped
       bool hasTapped = false;